mediacapture-streams/MediaStreamTrack-end-manual.https.html

<!doctype html>
<title>Test that mediastreamtrack are properly ended</title>
<link rel="author" title="Dominique Hazael-Massieux" href="mailto:dom@w3.org"/>
<link rel="help" href="http://w3c.github.io/mediacapture-main/getusermedia.html#mediastreamtrack">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<p class="instructions">When prompted, accept to share your video and audio
stream, and then revoke that permission.</p>
<h1 class="instructions">Description</h1>
<p class="instructions">This test checks that the video and audio tracks of
MediaStream object returned by the success callback in getUserMedia are
correctly set into inactive state when permission is revoked.</p>

<script>

promise_test(async t => {
  const stream = await navigator.mediaDevices.getUserMedia({audio: true, video: true});
  const vidTrack = stream.getVideoTracks()[0];
  assert_equals(vidTrack.readyState, "live", "The video track object is in live state");
  const audTrack = stream.getAudioTracks()[0];
  assert_equals(audTrack.readyState, "live", "The audio track object is in live state");

  const watcher = new EventWatcher(t, vidTrack, ["ended"]);
  await watcher.wait_for("ended");
  assert_equals(vidTrack.readyState, "ended", "Video track has been ended as expected");
  assert_equals(audTrack.readyState, "ended", "Audio track has been ended as expected");
  assert_false(stream.active, "MediaStream has been inactive as expected");
}, "Tests that the video MediaStreamTrack objects are properly ended on permission revocation");

</script>
